Move stubTokenProvider to testutil_test.go for explicit shared dependency
Extract the stubTokenProvider test helper from circuit_breaker_test.go into
a dedicated testutil_test.go file. This makes the cross-file dependency
between main_test.go and the shared helper explicit and discoverable,
following Go convention for test helpers used across multiple test files.

Closes #69
