+++ circuit_breaker_test.go
@@ -4,33 +4,11 @@ import (
 	"errors"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"testing"
 	"time"
 
 	gobreaker "github.com/sony/gobreaker/v2"
 )
-
-// stubTokenProvider is a controllable TokenProvider for testing.
-type stubTokenProvider struct {
-	calls  atomic.Int64
-	err    error  // when non-nil, GetToken returns this error
-	token  string // returned on success
-	closed atomic.Bool
-}
-
-func (s *stubTokenProvider) GetToken() (string, error) {
-	s.calls.Add(1)
-	if s.err != nil {
-		return "", s.err
-	}
-	return s.token, nil
-}
-
-func (s *stubTokenProvider) Close() error {
-	s.closed.Store(true)
-	return nil
-}
 
 // tripBreaker drives cbConsecutiveFailures failures through the provider to
 // open the circuit. The inner stub must already have a non-nil err.
